Run device-admin as `pi` user (#62)
This PR changes the device-admin ctr to run as `pi` instead of `root`,
so that USB drives manually mounted through device-admin's Storage
Drives page are mounted with `pi` ownership rather than `root` ownership
(as a follow-up to #60). This makes manual mounting in device-admin
behave consistently (in terms of ownership) with the behavior of
`usb-automount.pkg`.

To enable this change under a new `as-user-pi` feature flag in
`device-admin.pkg`, the feature flag exports polkit rules to so that
NetworkManager settings can be changed via the D-Bus API from the `pi`
user. The feature flag also exports a new systemd service to configure
tailscale to allow control from the `pi` user. Finally, this PR bumps
the `device-admin` ctr img to support these changes.

This work is tracked on Notion at
https://www.notion.so/Support-hotplugging-of-USB-drives-2a74e612c78a808c999ad81fb54a7c27?source=copy_link
.