Fix conflict between `tailscale funnel` and `forklift-apply.service` (#63)
@beniroquai had encountered a problem where, upon enabling [Tailscale
Funnel](https://tailscale.com/kb/1223/funnel) and rebooting the OS,
`forklift-apply.service` would fail because the Caddy ingress proxy
would fail to start. That failure occurs because Caddy tries to bind to
port 443 on all IP addresses, but when Tailscale Funnel is enabled it
binds to port 443 on a Tailscale-specific IP address. The fix for this
is to start Tailscale Funnel _after_ Caddy, so that Caddy won't
encounter that error. This PR implements that fix.

Note that this PR only fixes the conflict during boot; if the user
wishes to manually restart `forklift-apply.service`, they'll need to
first stop Tailscale Funnel.