Shrink the OS image (#51)
This PR follows up on a workaround added in
https://github.com/openUC2/rpi-imswitch-os/pull/48 for the GitHub
Actions CI runner running out of disk space in the pishrink step of the
CI workflow for building the OS image; that workaround restored disk
space margins by disabling the deployment of the offline docs site
originally added in https://github.com/openUC2/rpi-imswitch-os/pull/23,
but such a workaround is not a long-term solution.

Specifically, this PR removes the host installation of ImSwitch (which
uses up a lot of disk space and is only useful for modifying/testing
ImSwitch in software development, not for deployment by customers) from
the OS build scripts and instead allows a developer to opt-in to host
installation afterwards on an RPi with an internet connection, by
running the command `bash $(forklift plt locate-file
dx/host-imswitch.sh)` on the RPi.

This PR also:
- Bumps the RPi OS base from the version of bookworm built on 2025-05-13
to the version built on 2025-11-24
- Starts building a version of RPi OS based on trixie, for testing
purposes
- Stops trying to run `forklift stage apply` in the OS build system,
since it fails anyways
- Makes the OS image be built from the PR's HEAD commit rather than the
fictitious merge commit, as the latter makes the commit hash in
`forklift stage show` untraceable and requires `forklift plt upgrade` to
be run with `--force --allow-downgrade`

This work is tracked on Notion at
https://www.notion.so/Remove-unnecessary-modules-in-the-base-OS-build-scripts-2724e612c78a8083a953e0c890e6495c?source=copy_link
and
https://www.notion.so/Upgrade-OS-to-RPi-OS-13-trixie-29b4e612c78a8043a27fc27ed466ed04?source=copy_link
.