Prepopulate `boot-init-root`'s directory structure (#79)
Users trying to use the `boot-init-root` mechanism (as described in
https://www.notion.so/DN-25-Customer-specific-OS-pre-configuration-2924e612c78a80b9b97cfffc199f3ae5?v=28c4e612c78a804fa82b000c084384ed&source=copy_link#2924e612c78a8046ba8bf8c3a0fb7aa3
and as implemented in
https://github.com/openUC2/rpi-imswitch-os/tree/main/deployments/provisioning/boot-init-root.pkg)
need to create a bunch of nested directories themselves in order to
provision config files such as
`/home/pi/ImSwitchConfig/config/imcontrol_options.json`. However, we can
create those directories for them when we build the OS images. This PR
does exactly that.